fixing warning

+++ src/jet/live/LinkTimeRelocationsStep.cpp
@@ -70,9 +70,11 @@ namespace jet
                 continue;
             }
             if (reloc.size == sizeof(int32_t)) {
-                *reinterpret_cast<int32_t*>(relocAddress) += oldVar->runtimeAddress - relocSymbol->runtimeAddress;
+                *reinterpret_cast<int32_t*>(relocAddress) +=
+                    static_cast<int32_t>(oldVar->runtimeAddress) - static_cast<int32_t>(relocSymbol->runtimeAddress);
             } else if (reloc.size == sizeof(int64_t)) {
-                *reinterpret_cast<int64_t*>(relocAddress) += oldVar->runtimeAddress - relocSymbol->runtimeAddress;
+                *reinterpret_cast<int64_t*>(relocAddress) +=
+                    static_cast<int32_t>(oldVar->runtimeAddress) - static_cast<int32_t>(relocSymbol->runtimeAddress);
             }
             context->events->addLog(LogSeverity::kDebug, relocSymbol->name + " was relocated");
